requirements.txt
Flask==2.3.3
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
prophet==1.1.4
joblib==1.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.2
requests==2.31.0